@@ -1,4 +1,4 @@
-BioportalWebUi::Application.routes.draw do
+Rails.application.routes.draw do
 
   root :to => 'home#index'
 
